+++ src/layer/featureStoreLayerState.ts
@@ -3,8 +3,9 @@ export const featureStoreStateSymbol = Symbol('vcsFeatureType');
 /**
  * Enumeration of feature store item states
  */
-export enum FeatureStoreLayerState {
-  DYNAMIC = 'dynamic',
-  STATIC = 'static',
-  EDITED = 'edited',
-}
+export type FeatureStoreLayerState =
+  | 'dynamic'
+  | 'static'
+  | 'edited'
+  | 'deleted'
+  | 'removed';